- Fixed a bug where movies obtained from TMDB API had the parameters revenue or budget with a number bigger than the Int limit, so now both properties on MovieResponse class are Long. - Created MovieDAO class to prevent the same bug in case any movie has a movieId bigger than the Int limit, so now MovieDAO implements an UUID property managed by Room and the domain Movie class has a function to map itself to MovieDAO and vice-versa.

+++ app/src/main/java/com/simplepeople/watcha/data/model/MovieListResponse.kt
@@ -15,7 +15,7 @@ data class MovieListResponse(
         @SerializedName("adult") var adult: Boolean = false,
         @SerializedName("backdrop_path") var backdropPath: String = "",
         @SerializedName("genre_ids") var genreIds: ArrayList<Int> = arrayListOf(),
-        @SerializedName("id") var id: Int = 1,
+        @SerializedName("id") var id: Long = 1,
         @SerializedName("original_language") var originalLanguage: String = "",
         @SerializedName("original_title") var originalTitle: String = "",
         @SerializedName("overview") var overview: String = "",
